colab-requirements.txt
flwr[simulation]
wandb
nbformat
torchmetrics